requirements.txt
flask==3.0.3
Werkzeug==3.0.6
flask-cors==3.0.10
gunicorn==23.0.0
azure-ai-formrecognizer==3.3.3
azure-core==1.32.0
azure-storage-blob==12.19.0
numpy==2.1.2
pandas==2.2.3
python-dotenv==1.0.1
thefuzz==0.22.1
Flask-SQLAlchemy==3.1.1
flask-login==0.6.3
flask-mail==0.10.0
flask-migrate==4.1.0
psycopg2-binary==2.9.7
Pillow==10.1.0
pyodbc==5.2.0
sqlalchemy==2.0.40
sib-api-v3-sdk==7.6.0